+++ src/mem/size/bit.rs
@@ -194,8 +194,6 @@ bit_sized![= 8; for crate::char8];
 bit_sized![= 16; for crate::char16];
 #[cfg(feature = "_char24")]
 bit_sized![= 24; for crate::char24];
-#[cfg(feature = "_char32")]
-bit_sized![= 32; for crate::char32];
 
 bit_sized![= 8; for NonZeroI8, NonZeroU8];
 bit_sized![= 16; for NonZeroI16, NonZeroU16];