+++ src/text/char/core_impls.rs
@@ -65,8 +65,7 @@ core_impls![
     char7 | "_char7" + Self(unwrap![some NonExtremeU8::new(0)]),
     char8 | "_char8" + Self(0),
     char16 | "_char16" + Self(unwrap![some NonSurrogateU16::new(0)]),
-    char24 | "_char24" + Self { hi: unwrap![some NonExtremeU8::new(0)], mi: 0, lo: 0 },
-    char32 | "_char32" + Self('\x00')
+    char24 | "_char24" + Self { hi: unwrap![some NonExtremeU8::new(0)], mi: 0, lo: 0 }
 ];
 
 /* From char7 */
@@ -100,13 +99,6 @@ mod c7 {
         #[must_use]
         fn from(c: char7) -> char24 {
             c.to_char24()
-        }
-    }
-    #[cfg(feature = "_char32")]
-    impl From<char7> for char32 {
-        #[must_use]
-        fn from(c: char7) -> char32 {
-            c.to_char32()
         }
     }
 }
@@ -144,13 +136,6 @@ mod c8 {
             c.to_char24()
         }
     }
-    #[cfg(feature = "_char32")]
-    impl From<char8> for char32 {
-        #[must_use]
-        fn from(c: char8) -> char32 {
-            c.to_char32()
-        }
-    }
 }
 
 /* From char16 */
@@ -184,20 +169,6 @@ mod c16 {
         #[must_use]
         fn from(c: char16) -> char24 {
             c.to_char24()
-        }
-    }
-    #[cfg(feature = "_char32")]
-    impl From<char16> for char32 {
-        /// # Features
-        /// Makes use of the `unsafe_str` feature if enabled.
-        #[must_use]
-        fn from(c: char16) -> char32 {
-            #[cfg(any(feature = "safe_text", not(feature = "unsafe_str")))]
-            return c.to_char32();
-
-            #[cfg(all(not(feature = "safe_text"), feature = "unsafe_str"))]
-            // SAFETY: we've already checked we contain a valid char.
-            return unsafe { char32(char::from_u32_unchecked(c.0.get() as u32)) };
         }
     }
 }
@@ -235,62 +206,6 @@ mod c24 {
             c.try_to_char16()
         }
     }
-    #[cfg(feature = "_char32")]
-    impl From<char24> for char32 {
-        /// # Features
-        /// Makes use of the `unsafe_str` feature if enabled.
-        #[must_use]
-        fn from(c: char24) -> char32 {
-            #[cfg(any(feature = "safe_text", not(feature = "unsafe_str")))]
-            return c.to_char32();
-
-            #[cfg(all(not(feature = "safe_text"), feature = "unsafe_str"))]
-            // SAFETY: we've already checked we contain a valid char.
-            return unsafe { char32(char::from_u32_unchecked(c.to_u32())) };
-        }
-    }
-}
-
-/* From char32 */
-
-#[cfg(feature = "_char32")]
-mod c32 {
-    use super::*;
-
-    impl From<char32> for char {
-        #[must_use]
-        fn from(c: char32) -> char {
-            c.to_char()
-        }
-    }
-    #[cfg(feature = "_char7")]
-    impl TryFrom<char32> for char7 {
-        type Error = TextError;
-        fn try_from(c: char32) -> Result<char7> {
-            c.try_to_char7()
-        }
-    }
-    #[cfg(feature = "_char8")]
-    impl TryFrom<char32> for char8 {
-        type Error = TextError;
-        fn try_from(c: char32) -> Result<char8> {
-            c.try_to_char8()
-        }
-    }
-    #[cfg(feature = "_char16")]
-    impl TryFrom<char32> for char16 {
-        type Error = TextError;
-        fn try_from(c: char32) -> Result<char16> {
-            c.try_to_char16()
-        }
-    }
-    #[cfg(feature = "_char24")]
-    impl From<char32> for char24 {
-        #[must_use]
-        fn from(c: char32) -> char24 {
-            c.to_char24()
-        }
-    }
 }
 
 /* From char */
@@ -321,12 +236,5 @@ impl From<char> for char24 {
     #[must_use]
     fn from(c: char) -> char24 {
         char24::from_char(c)
-    }
-}
-#[cfg(feature = "_char32")]
-impl From<char> for char32 {
-    #[must_use]
-    fn from(c: char) -> char32 {
-        char32::from_char(c)
     }
 }