+++ src/text/char/impls/c8.rs
@@ -54,16 +54,6 @@ impl char8 {
             Err(CharConversion)
         }
     }
-    /// Tries to convert a `char32` to `char8`.
-    #[cfg(feature = "_char32")]
-    #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "_char32")))]
-    pub const fn try_from_char32(c: char32) -> Result<char8> {
-        if Char::byte_len(c.to_u32()) == 1 {
-            Ok(char8(c.to_u32() as u8))
-        } else {
-            Err(CharConversion)
-        }
-    }
     /// Tries to convert a `char` to `char8`.
     pub const fn try_from_char(c: char) -> Result<char8> {
         if Char::byte_len(c as u32) == 1 {
@@ -113,13 +103,6 @@ impl char8 {
     #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "_char24")))]
     pub const fn to_char24(self) -> char24 {
         char24::from_char8(self)
-    }
-    /// Converts this `char8` to `char32`.
-    #[must_use]
-    #[cfg(feature = "_char32")]
-    #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "_char32")))]
-    pub const fn to_char32(self) -> char32 {
-        char32::from_char8(self)
     }
     /// Converts this `char8` to `char`.
     #[must_use]