+++ src/text/grapheme/string.rs
@@ -64,23 +64,6 @@ impl GraphemeString {
         Str::from_utf8(&c.to_utf8_bytes()).unwrap().into()
     }
 
-    /// Creates a new `GraphemeString` from a `char32`.
-    /// # Features
-    /// Makes use of the `unsafe_str` feature if enabled.
-    #[must_use]
-    #[cfg(feature = "_char32")]
-    #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "_char32")))]
-    #[cfg(feature = "dep_unicode_segmentation")]
-    #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "dep_unicode_segmentation")))]
-    pub fn from_char32(c: char32) -> GraphemeString {
-        #[cfg(any(feature = "safe_text", not(feature = "unsafe_str")))]
-        return Str::from_utf8(&c.to_utf8_bytes()).unwrap().into();
-        #[cfg(all(not(feature = "safe_text"), feature = "unsafe_str"))]
-        unsafe {
-            Str::from_utf8_unchecked(&c.to_utf8_bytes()).into()
-        }
-    }
-
     /// Creates a new `GraphemeString` from a `char`.
     /// # Features
     /// Makes use of the `unsafe_str` feature if enabled.