+++ src/text/str/string_u.rs
@@ -63,8 +63,7 @@ macro_rules! impl_string_u {
         ///     [`7`](Self::from_char7),
         ///     [`8`](Self::from_char8),
         ///     [`16`](Self::from_char16),
-        ///     [`24`](Self::from_char24),
-        ///     [`32`](Self::from_char32)
+        ///     [`24`](Self::from_char24).
         ///   )*.
         /// - Deconstruct:
         ///   [`into_array`][Self::into_array],
@@ -428,19 +427,6 @@ macro_rules! impl_string_u {
                 if new.len > 2 { new.arr[2] = bytes[2]; }
                 if new.len > 3 { new.arr[3] = bytes[3]; }
                 Ok(new)
-            }
-
-            #[doc = "Creates a new `String" $t:camel "` from a `char32`."]
-            ///
-            /// # Panics
-            #[doc = "Panics if `CAP > `[`" $t
-                "::MAX`]` || CAP < c.`[`len_utf8()`][char32#method.len_utf8]."]
-            ///
-            #[doc = "Will never panic if `CAP >= 4 && CAP <= `[`" $t "::MAX`]."]
-            #[cfg(feature = "_char32")]
-            #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "_char32")))]
-            pub const fn from_char32(c: char32) -> Result<Self> {
-                Ok(unwrap![ok? Self::from_char(c.0)])
             }
 
             /* from bytes */